% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addpart.R
\name{addpart}
\alias{addpart}
\title{Additive Partitioning}
\usage{
addpart(d, depmix, name.CE = NULL, name.SE = NULL, excl.zeroes = FALSE)
}
\arguments{
\item{d}{Data frame describing the composition of plant
communities, on a per-species basis}

\item{depmix}{A 'model formula' describing the structure of the
data, in the \code{form y ~ mixture/species + plot}}

\item{name.CE}{Name of column holding the calculated complemetarity
effect. Defaults to \code{CE.y}, where \code{y} is replaced by the
name of the dependent variable (usually the biomass measure).}

\item{name.SE}{Name of column holding the calculated selection
effect. Defaults to \code{SE.y}, where \code{y} is replaced by the
name of the dependent variable specified.}

\item{excl.zeroes}{Exclude species with unknown or zero biomass
from calculations.}
}
\value{
A data frame containing columns for plant species mixture,
plot, and the computed complementarity and selection effects.
}
\description{
This function computes Loreau and Hectors additive partitioning,
given a data set that specifies the composition of plant
communities on a species basis.
}
\details{
The input data frame needs to contain columns defining the units
for which complementarity and selection effects are to be
calculated (typically field plots), the community composition of
these plots, the species contained in these plots, plus their
contribution to the community-level metric that is to be analysed
(typically biomass). These columns are specified using a model
formula of the form \code{form y ~ mixture/species + plot}. Note
that plot and mixture can be identical, in which case one has to
write \code{form y ~ mixture/species + mixture}.

Species composition is a character string constructed so that the
presence of individual species can be tested by searching for the
occurence of the species code in the composition string. For
example, species \code{A} will be found in composition \code{ABCD}
but not in \code{EFGH}. The composition codes can contain
additional characters as long as the above rule is met,
e.g. \code{A|B|C|D} would work as well.

Monoculture biomass (the reference for the underlying relative
yield calculation) is calculated as mean of all the monoculture
plots.

Relative yields, on which the additive partitioning scheme bases,
cannot be calculated for communities that contain species with zero
or unknown monoculture biomass. For such plots, the complementarity
and selection effects will be set to NA, unless
\code{excl.zeroes=TRUE} is specified. Then, these species will be
removed from the species set before complementarity and selection
effects are calculated.
}
\examples{
## TO BE ADDED
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\references{
Loreau M \& Hector A (2001) Partitioning selection and
complementarity in biodiversity experiments. Nature 412, 72-76
}

