% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcFD.R
\name{calcFD}
\alias{calcFD}
\title{Calculate Petchey \& Gaston's functional diversity index}
\usage{
calcFD(communities = NULL, traits = NULL, distance = "euclidean",
  method = "complete", format = "default", which = c("FD"),
  shrink.tree = TRUE)
}
\arguments{
\item{communities}{data frame containing a column with species
names and a column designating the communities (e.g. plots or other groups to compare)}

\item{traits}{data frame containing the traits of each species}

\item{distance}{the method used to calculate distances in trait
space. See \code{\link{dist}} for details.}

\item{method}{the agglomeration method to be used. See
\code{\link{hclust}} for details.}

\item{format}{indicates in which format community composition is
passed. Can be either \code{default} or \code{matrix}}

\item{which}{indicates which results to return
(\code{default="FD"}). The vector may contain any combination of
\code{"FD"},\code{"stdFD"}, \code{"maxFD"}, \code{"distances"},
\code{"tree"}, \code{"branches"}. \code{"FD"} indicates functional
diversity (i.e. total branch length of subtree), \code{"stdFD"}
standardized functional diversity (branch length divided by total
tree branch length) and \code{"maxFD"} total tree branch
length. \code{"distances"} also returns the calculated distance matrix,
\code{"tree"} the calculated trait tree, and \code{"branches"} a list of branch
indiced within the tree that connect a tip to the root.}

\item{shrink.tree}{logical indicating whether species not present
in any community should be removed from the trait tree. defaults
to \code{TRUE}.}
}
\value{
a list containing the items selected by \code{which}.
}
\description{
Calculates functional diversity sensu Petchey & Gaston, including
the changes suggested in their 2006 update.
}
\details{
Note that this function converts the tree obtained from the
hierarchical cluster analysis (\code{hcluse}) to a tree of class
\code{phylo}, using \code{as.phylo}.  This introduces a scaling
factor of two. As a result, the absolute measure \code{FD} is only
half the one obtained by Owen Petchey and Jens Schumacher's
code. Standardized FD is equal, though.
}
\examples{
library(pdiv)

data(composition.example)
composition.example
##    community species
## 1          1       a
## 2          1       b
## 3          1       c
## ...
## 19         7       c
## 20         8       d

data(traits.example)
traits.example
##    species     trait1    trait2     trait3
## 1        a 0.39486189 0.9727755 0.83947128
## 2        b 0.37285983 0.6587653 0.94541045
## 3        c 0.06773421 0.0873821 0.39090723
## ...
## 9        i 0.95826170 0.1678766 0.92731325
## 10       j 0.29947903 0.3547652 0.36046854

calcFD(composition.example,traits.example)
## $FD
##    1    2    3    4    5    6    7    8
## 1.63 1.25 1.46 2.10 0.00 0.00 0.00 0.00
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\references{
Petchey O \& Gaston K (2006) Functional diversity: back to basics and looking forward.
Ecology Letters 9: 741--758.

Petchey OL \& Gaston KJ (2002) Functional diversity (FD), species
richness and community composition. Ecology Letters 5: 402--411.
}
\seealso{
\code{\link{calcPD}}
}

