% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMPD.R
\name{calcMPD}
\alias{calcMPD}
\title{Calculate phylogenetic diversity as mean phylogenetic distance of pairs}
\usage{
calcMPD(communities = NULL, tree = NULL, which = c("MPD"),
  format = "default", scale = 0.5, abundance = FALSE)
}
\arguments{
\item{communities}{data frame containing a column specifying the
communities (e.g. plots or other groups to compare), a column with
species names, and optionally a column with species abundances.
Species names must match the tip names in the phylogenetic tree.
The abundances column is only used if \code{abundance = TRUE}.}

\item{tree}{phylogenetic tree of class \code{phylo}}

\item{which}{indicates which results to return
(\code{default="MPD"}). The vector may contain 
\code{"MPD"}, and \code{"dist"}. \code{"MPD"}
indicates mean phylogenetic diversity (i.e. mean age of last common ancestor),
\code{"dist"} will return the age of the last common ancestor for all pairs.}

\item{format}{indicates in which format community composition is
passed. Can be either \code{default} or \code{matrix}.}

\item{scale}{a user-defined scale factor (default 0.5) by which the
total branch length of the connection between a pair is
scaled. This is useful in ultrametric trees to arrive at the
distance to the last common ancestor since the branch length from
both tips would be cumulated otherwise.}

\item{abundance}{logical indicating whether abundances are used
to weigh the distance between each pair.}
}
\value{
a list containing the items selected by \code{which}.
}
\description{
Given a phylogenetic tree and data on community composition, the
average distance between all species pairs within a community is
calculated.  This procedure optionally includes abundance-weighing
(weights are the product of the abundances of the species of each pair).
The distance metric used is the sum of the branch lengths from the
two species' tips up to the last common ancestor, multiplied by a
user-defined scale factor.
}
\details{
On platforms other than Windows, \code{calcMPD} parallelizes the
time-critical steps if the \code{parallel} package is
available. Parallelization is not available on Windows because of a
lack of \code{fork} functionality on which \code{mclapply} bases.
}
\examples{

library(pdiv)
data(composition.example)
data(phylotree.example)

plot(phylotree.example)

calcMPD(composition.example,phylotree.example)
## $MPD
## [1] 333  85 293 308   0   0   0   0

ex2 <- composition.example
ex2$abu<-runif(nrow(composition.example))
calcMPD(ex2,phylotree.example,which="MPD",abundance=TRUE)

calcMPD(composition.example,phylotree.example,which="dist")
## $dist
## a|j a|b a|c a|d a|e a|f a|h b|c b|d b|e b|f b|h c|d
## 385 295 295 385 385 385 385 250 385 385 385 385 385
## f|j f|h h|j c|e c|f c|h d|e d|h e|j e|h
## 109 109  37 385 385 385  97 109 109 109

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\references{
Warwick RM and Clark KR (1995) New 'biodiversity' measures
reveal a decrease in taxonomic distinctness with increasing stress.
Mar Ecol Prog Ser 29:301-305.
}
\seealso{
\code{\link{calcPD}}

\code{\link{communityToMatrix}}

\code{\link{communityFromMatrix}}
}

