% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcPD.R
\name{calcPD}
\alias{calcPD}
\title{Calculate phylogenetic diversity}
\usage{
calcPD(communities = NULL, tree = NULL, which = c("PD"),
  format = "default")
}
\arguments{
\item{communities}{data frame containing a column with species
names and a column designating the communities (e.g. plots or other
groups to compare)}

\item{tree}{phylogenetic tree of class \code{phylo}}

\item{which}{indicates which results to return
(\code{default="PD"}). The vector may contain any combination of
\code{"PD"}, \code{"maxPD"}, and \code{"branches"}. \code{"PD"}
indicates phylogenetic diversity (i.e. total branch length of
subtree), \code{"maxPD"} total tree branch length, and
\code{"branches"} a list of branch indiced within the tree that
connect a tip to the root.}

\item{format}{indicates in which format community composition is
passed. Can be either \code{default} or \code{matrix}}
}
\value{
a list containing the items selected by \code{which}.
}
\description{
There are also other routines to calculate phylogenetic
diversity. I implemented this function to have one that follows
exactly the same procedure as the metric defined by Petchey &
Gaston (see \code{\link{calcFD}}). Phylogenetic diversity is
calculated as total branch length of the phylogenetic subtree
defined by the set of species present in a community.
}
\examples{
library(pdiv)
data(composition.example)
data(phylotree.example)
plot(phylotree.example)
calcPD(composition.example,phylotree.example)
## $PD
##   com        PD
## 1   1 1315.1436
## 2   2  254.8177
## 3   3  878.9794
## 4   4 1520.9276
## 5   5    0.0000
## 6   6    0.0000
## 7   7    0.0000
## 8   8    0.0000
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\seealso{
\code{\link{calcFD}}
}

