% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{calcTp}
\alias{calcTp}
\title{Calculate transitivity from transition matrix}
\usage{
calcTp(P)
}
\arguments{
\item{P}{square P matrix}
}
\value{
transitivity measure `tau_P'
}
\description{
Given a transition matrix P, this function calculates transitivity tau_P as defined
in equation 19 in Ulrich et al. (2014).
}
\examples{
library(pdiv)
## example from Fig. 2 in Ulrich et al.
C1<-matrix(c(1,1,1,0,1,1,0,0,1),byrow=TRUE,nrow=3)
C4<-matrix(c(1,.8,.3,.2,1,.8,.7,.2,1),byrow=TRUE,nrow=3)
P1<-calcPfromC(C1)
P4<-calcPfromC(C4)
calcTp(P1)
calcTp(P4) 
}
\references{
Ulrich et al. (2014) Matrix models for quantifying
    competitive intransitivity from species abundance data. Oikos
    123:1057-1070.
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
