% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdiv-data.R
\docType{data}
\name{composition.example}
\alias{composition.example}
\title{Made-up example to demonstrate calcFD, calcPD, communityToMatrix}
\format{
\code{data.frame}
}
\description{
table with 8 communities and their species
}
\author{
Pascal Niklaus
}
