% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdiv-data.R
\docType{data}
\name{divpart}
\alias{divpart}
\title{Made-up example to demonstrate addpart and tripart}
\format{
object of class \code{data.frame}
}
\description{
Biomass of species in plots differing in composition. Used to demonstrate partitioning schemes
for net biodiversity effects (see \code{addpart} and \code{tripart}).
}
\author{
Pascal Niklaus
}
