% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hill.R
\name{hill.Na}
\alias{evenness.1overD}
\alias{hill.Ha}
\alias{hill.Na}
\alias{shannon.H}
\alias{simpson.D}
\title{Hill's series of diversity indices}
\usage{
hill.Na(x, a = 1)

hill.Ha(x, a = 1)

shannon.H(x)

simpson.D(x)

evenness.1overD(x)
}
\arguments{
\item{x}{Proportional measure of species abundances. These need not
be standardized to unity sum.}

\item{a}{Order of diversity number to be calculated. Defaults to \code{1}.}
}
\value{
Diversity metric
}
\description{
Hill proposed a unified notation of diversity numbers of different
orders \code{a}. As special cases, it also contains the number of
species (\code{a=0}), Shannon's index of diversity (\code{a=1}),
and the inverse of Simpson's index (\code{a=2}). \code{a=-Inf} and
\code{a=+Inf} correspond to the inverse of the proportional
abundance of the rarest and the most common species.
}
\details{
\code{hill.Na(x,a)} calculated these diversity numbers, whereas
\code{hill.Ha(x,a)} computes the generalized entropy measures of
order \code{a}.  For convenience, \code{simpson.D} and
\code{shannon.H} define these special cases and essentially yield
\code{1/hill.Na(x,2)} and \code{hill.Ha(x,1)}.

Many measures of diversity indices have been proposed, which have
different mathematical properties. A particularly useful one is
implemented as \code{evenness.1overD}, and returns \code{1/(DS)},
where D is Simpson's index of dominance and S is the total number
of species present. It has many desirable properties, including
independence from species richness, which is not the case for many
other metrics (see Smith and Wilson for a detailed illustration,
and example below).
}
\examples{
library(pdiv)

x1 <- c(1,1,1,3)
x2 <- rep(x1,2)

shannon.H(x1)
## [1] 1.242453

simpson.D(x1)
## [1] 0.3333333

hill.Na(x1)
## [1] 3.464102
hill.Ha(x1,1)
## [1] 1.242453               # equal to Shannon's H

hill.Na(x1,0)
## [1] 4                      # species numbers

hill.Na(x1,2)
## [1] 3                      # equal to 1/D

evenness.1overD(x1)           # note that E(1/D) is
## [1] 0.75                   # equal for x1 and x2
evenness.1overD(x2)
## [1] 0.75

shannon.H(x1)/log(length(x1)) # whereas this is
## [1] 0.8962406              # not the case for
shannon.H(x2)/log(length(x2)) # the Shannon-based
## [1] 0.9308271              # evenness index
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\references{
Hill MO (1973) Diversity and evenness: a unifying
notation and its consequences. Ecology 54(2):428--432.

Smith B & Wilson JB (1996) A consumer's guide to
diversity indices, OIKOS 76: 70--82.
}

