% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levinsB.R
\name{levinsB}
\alias{levinsB}
\title{Levins' B}
\usage{
levinsB(x)
}
\arguments{
\item{x}{Measure of species abundance in niches. These need not
be standardized to unity sum.}
}
\value{
Niche breadth as defined by Levins.
}
\description{
Calculate niche breadth as defined by Levins (1968).
}
\examples{
library(pdiv)
x <- c(0,1,2,3,2,0)
levinsB(x)

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\references{
Levins (1968) Evolution in changing
    environments. Princeton University Press, Princeton, New
    Jersey, USA.
}

