% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdiv-data.R
\docType{data}
\name{phylotree.example}
\alias{phylotree.example}
\title{Made-up example to demonstrate calcPD}
\format{object of class \code{phylo}}
\description{
phylogenetic tree with 20 species
}
\author{
Pascal Niklaus
}

