% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.sim.R
\name{prop.sim}
\alias{prop.sim}
\title{Proportional similarity}
\usage{
prop.sim(x, y)
}
\arguments{
\item{x}{Measure of species abundance in niches. These need not
be standardized to unity sum.}

\item{y}{Similar to x but for 2nd species}
}
\value{
Proportional similarity index.
}
\description{
Calculate proportional similarity as defined in Colwell and Futuyma (1971).
}
\examples{
library(pdiv)
x <- c(0,0,1,2,1,0)
y <- c(0,2,1,0,0,0)
prop.sim(x,y)

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\references{
Colwell and Futuyma (1971) Measurement of niche breadth
    and overlap. Ecology 52:567--576.
}

