% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdiv-data.R
\docType{data}
\name{traits.example}
\alias{traits.example}
\title{Made-up example to demonstrate calcFD}
\format{\code{data.frame}}
\description{
table with 10 species and three traits per species
}
\author{
Pascal Niklaus
}

